requirements.txt
# YouTube Shorts Automation - Python Dependencies
# Pin versions for reproducibility

# LLM API client (OpenRouter-compatible OpenAI SDK)
openai==1.30.1

# Text-to-Speech (no API key required)
gTTS==2.5.1

# HTTP requests (Pexels API, downloads)
requests==2.31.0

# Video editing
moviepy==1.0.3
imageio==2.34.0
imageio-ffmpeg==0.4.9
Pillow==10.3.0

# Google / YouTube API
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
httplib2==0.22.0

# Environment variable support
python-dotenv==1.0.1

# Retry / resilience
tenacity==8.3.0

# Utilities
tqdm==4.66.4
